007-03 gemini_asst_revised/007-04 gemini_asst_integrated/langgraph_agents/requirements.txt
langchain>=0.1.0
langgraph>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.1
discord.py>=2.3.0
notion-client>=2.2.1
google-api-python-client>=2.95.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0